==> 2023/2.add_nest_parent_.ipynb <==
import pandas as pd

df = pd.read_excel('removed_df.xlsx')

df['temp'] = df['hs'].fillna(method='ffill')
df['temp2'] = df['temp'].shift(1)

def get_nest_parent_(row):
    hs = str(row['hs'])
    if hs == 'nan':
        if len(str(row['temp2'])) > 5:
            return str(row['temp2'])[:2] + "." + str(row['temp2'])[2:4]
        return row['temp2']
    
    if len(hs) > 5:
        return hs[:2] + "." + hs[2:4]

    return int(hs[:2])

df['_nest_parent_'] = df.apply(get_nest_parent_, axis=1)
cols = df.columns.to_list()
cols = cols[:5] + ['_nest_parent_'] + cols[5:-4]

df[cols].to_excel('has_nest_parent.xlsx', index=False)
